extracting DropDownList and IntervalDropDown as separate components
